Fix TCG alignment problems on Sparc64 host
git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@4485 c046a42c-6fe2-441c-8c8c-71466251a162

+++ tcg/tcg.h
@@ -71,7 +71,7 @@ typedef struct TCGRelocation {
 } TCGRelocation; 
 
 typedef struct TCGLabel {
-    int has_value;
+    long has_value; // long instead of int to enforce alignment
     union {
         tcg_target_ulong value;
         TCGRelocation *first_reloc;
@@ -80,7 +80,7 @@ typedef struct TCGLabel {
 
 typedef struct TCGPool {
     struct TCGPool *next;
-    int size;
+    long size; // long instead of int to enforce alignment
     uint8_t data[0];
 } TCGPool;
